Contract Test 추가
