+++ ticketIssuance/src/main/java/ticketIssuance/TicketIssuanceController.java
@@ -1,16 +1,25 @@
 package ticketIssuance;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
  @RestController
  public class TicketIssuanceController {
+
+  @PutMapping("/ticketIssuances/cancel/{bookId}")
+  public TicketIssuance payCanceled(@PathVariable("bookId") Long bookId) {
+
+   TicketIssuance ti = new TicketIssuance();
+   ti.setBookId(bookId);
+   ti.setIssueStatus("Canceled");
+
+   return ti;
+
+  }
